+++ CHANGELOG.md
@@ -1,7 +1,12 @@
+### 0.3.0
+
+*   add `successProps` and `errorProps` props
+*   add callback props: `onPending`, `onFinished`, `onSuccess`, `onSuccessEnd`, `onError`, `onErrorEnd`
+
 ### 0.0.7
 
-- add `initialPending` prop
+*   add `initialPending` prop
 
 ### 0.2.0
 
-- Handle resolved promises as failures when their payload is of type `Error`
+*   Handle resolved promises as failures when their payload is of type `Error`